treewide: remove susfs4ksu due to syscall issues
for now, susfs is a "working" way as a root hiding method,
however it's a dirty way to do.

1. syscall speed was really hurt by this stuff,
as a test, average hackbench result with legacy root (magisk) is around 1.8s, and 1.9s with ksu.
now with a fucking susfs, it spikes up to 2.1s which should be a concern.

2. as we can see in dmesg, we got a lot of susfs mount and unmount operation which cause latency and jitter to spike.

3. another shit i have that it fucking breaks dirty builds for additional susfs stuff, which really hurts compile time and fucking cook my laptop since i always do full lto builds.

so fuck this, and i don't give a fuck about your failing root detection.

Signed-off-by: kvsnr113 <kvsnrprojkt113@gmail.com>