treewide: introduce E404 kernel signature build
e404_attributes is a kernel module to control kernel behaviour,
currently working on a sysfs and cmdline to store data.
this provides functional stuff to the userspace while keeping it simple.

i'm already kind enough to make this open source so people could see what i am doing,
kangers will be happy to see this tho.

Signed-off-by: kvsnr113 <kvsnrprojkt113@gmail.com>